# tidy: check for invalid file names