@Kobzol 2025-07-15
Just out of curiosity, how long does it take to run this on your machine?