@samueltardieu 2025-07-15
To run `git ls-files` in my working directory, around 16ms wall time. Timing `x test tidy` gives no perceivable difference in walltime (4.22s vs. 4.26s, but it can vary in both directions – 24 virtual cores laptop):

```txt
with this PR:

Executed in    4.26 secs    fish           external
   usr time   14.15 secs    3.70 millis   14.15 secs
   sys time    5.47 secs    1.12 millis    5.47 secs

without this PR:

Executed in    4.22 secs    fish           external
   usr time   14.35 secs    0.79 millis   14.35 secs
   sys time    5.40 secs    1.13 millis    5.40 secs
```