@Kobzol 2025-07-15
Okay, that should be fine. Hopefully it won't be sluggish e.g. on Windows, but I think that tidy does so much I/O anyway that this won't be bad.