@samueltardieu 2025-07-15
Yeah, since it uses only file names and not even file contents, there is a good chance that the cache is primed already (or that this primes it for other parts of tidy). If it gets r+, maybe rollup=never would make it easier to roll it back in case it is heavier than expected on some platforms.